search: move date visualization in search-date component
