+++ app/components/search-date.html
@@ -1,0 +1,7 @@
+<h5>Date</h5>
+<chartist class="ct-chart ph-search-chart-date"
+  chartist-data="$ctrl.chartistData"
+  chartist-chart-options="$ctrl.chartistOptions"
+  chartist-chart-type="Bar"
+  chartist-events="$ctrl.chartistEvents"
+></chartist>
