+++ app/components/search.html
@@ -27,13 +27,9 @@
   <div class="container">
     <div class="row">
       <div class="col-md-3 text-center">
-        <h5>Date</h5>
-        <chartist class="ct-chart ph-search-chart-date"
-          chartist-data="$ctrl.chartistData.date"
-          chartist-chart-options="$ctrl.chartistOptions"
-          chartist-chart-type="Bar"
-          chartist-events="$ctrl.chartistEvents"
-        ></chartist>
+        <search-date
+          facet="$ctrl.facets.publishedByYear"
+        ></search-date>
       </div>
       <div class="col-md-3 text-center">
         <h5>Journal</h5>
